Pure EditorScreen works in storybook
